requirements.txt
Django==2.2.3
django-classic-user-accounts==1.0.37
djangorestframework==3.10.1
djangorestframework-simplejwt==4.3.0
Pillow==6.1.0
PyJWT==1.7.1
pytz==2019.1
sorl-thumbnail==12.5.0
sqlparse==0.3.0
